hidl: fingerprint: Handle vendor capture events for optical FODs
It seems that Samsung optical FODs use different, non-AOSP codes to represent sensor events. Testing shows that the vendor fingerprint library on the Galaxy A50 doesn't use standard replies for the onFingerDown and onFingerUp to work on its own. The older Lineage 18.1 FOD driver captures vendor events to make this possible.

Furthermore, it also sends additional vendor requests to do authentication when the sensor is pressed. Add that as well.

Co-authored-by: santhosh6194 <santhosh6194@gmail.com>
Signed-off-by: John Vincent <git@tenseventyseven.cf>