hidl: fingerprint: Send `fod_enable` on fingerDown and on fingerUp
Most ROMs have reverted the on*UdfpsOverlay callbacks, use the one we're using for FingerprintInscreen.

While we're at it, fix compilation issues, and move HBM commands back to onFingerDown and onFingerUp (these are called by Android, not by the HAL. We may consider putting TSP and additional request commands to here as well.

Signed-off-by: John Vincent <git@tenseventyseven.cf>